backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-socketio==5.11.0
pymetasploit3==1.0.3
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.6
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic-settings==2.1.0
httpx==0.26.0
websockets==12.0
